+++ lib/gvc/gvc.c
@@ -25,9 +25,9 @@ GVC_t *gvContext(void)
 {
     GVC_t *gvc;
 
-    agattr_text(NULL, AGNODE, "label", NODENAME_ESC);
     /* default to no builtins, demand loading enabled */
     gvc = gvNEWcontext(NULL, true);
+    gvconfig_default_attrs(gvc);
     gvconfig(gvc, false); /* configure for available plugins */
     return gvc;
 }
@@ -36,8 +36,8 @@ GVC_t *gvContextPlugins(const lt_symlist_t *builtins, int demand_loading)
 {
     GVC_t *gvc;
 
-    agattr_text(NULL, AGNODE, "label", NODENAME_ESC);
     gvc = gvNEWcontext(builtins, demand_loading);
+    gvconfig_default_attrs(gvc);
     gvconfig(gvc, false); /* configure for available plugins */
     return gvc;
 }